change double to float, make null definition of animal

+++ src/ensure/animal/amphibi/AfricanFrog.java
@@ -6,12 +6,12 @@ package ensure.animal.amphibi;
  */
 public class AfricanFrog extends Amphibi {
     //methods
-    public AfricanFrog(String inputName, float inputWeight, boolean isTamed, int x, int y, String inputFoodType, String inputHabitat){
+    public AfricanFrog(String inputName, double inputWeight, boolean isTamed, int x, int y, String inputFoodType, String inputHabitat){
         super(inputName,inputWeight,isTamed,x,y,inputFoodType,inputHabitat);
     }
 
-    public float getTotalFood() {
-        return (float) (super.getWeight() * 0.01);
+    public double getTotalFood() {
+        return (double) (super.getWeight() * 0.01);
     }
     public void interact(){
         System.out.println("Kroakk!");
