+++ src/ensure/animal/mamalia/Tiger.java
@@ -6,12 +6,12 @@ package ensure.animal.mamalia;
  */
 public class Tiger extends Mamalia {
     //methods
-    public Tiger(String inputName, float inputWeight, boolean isTamed, int x, int y, String inputFoodType, String inputHabitat){
+    public Tiger(String inputName, double inputWeight, boolean isTamed, int x, int y, String inputFoodType, String inputHabitat){
         super(inputName,inputWeight,isTamed,x,y,inputFoodType,inputHabitat);
     }
 
-    public float getTotalFood(){
-        return (float) (super.getWeight() * 0.1);
+    public double getTotalFood(){
+        return (double) (super.getWeight() * 0.1);
     }
     public void interact(){
         System.out.println("Growl!");
