+++ src/ensure/animal/pisces/Pisces.java
@@ -8,7 +8,7 @@ import ensure.animal.Animal;
  */
 abstract class Pisces extends Animal {
     //methods
-    public Pisces(String inputName, float inputWeight, boolean isTamed, int x, int y, String inputFoodType, String inputHabitat){
+    public Pisces(String inputName, double inputWeight, boolean isTamed, int x, int y, String inputFoodType, String inputHabitat){
         super(inputName,inputWeight,isTamed,x,y,inputFoodType,inputHabitat);
     }
